move plot into mod-sys-analyser
